fixed methods for User statistics

+++ src/main/java/com/softserve/academy/dao/UserDao.java
@@ -4,6 +4,6 @@ import com.softserve.academy.entity.User;
 
 public interface UserDao {
     User getUserById(Integer id);
-    double getUserStatisticAverageAge();
-    double getUserAverageTimeOfUsingLibrary();
+    int getUserStatisticAverageAge();
+    int getUserAverageTimeOfUsingLibrary();
 }
