+++ src/main/java/com/softserve/academy/service/UserService.java
@@ -5,7 +5,7 @@ import com.softserve.academy.entity.User;
 public interface UserService {
     User getUserById(Integer id) throws IllegalArgumentException;
 
-    double getUserStatisticAverageAge();
+    int getUserStatisticAverageAge();
 
-    double getUserAverageTimeOfUsingLibrary();
+    int getUserAverageTimeOfUsingLibrary();
 }
